Revert "avoid long -> float"
This reverts commit 37ad226a3d6b934bbefcdc0927d23c33e4305f61.

+++ firmware/controllers/algo/dynoview.cpp
@@ -69,7 +69,7 @@ void DynoView::update(vssSrc src) {
 void DynoView::updateAcceleration(efitimeus_t deltaTime, float deltaSpeed) {
     // todo: explain why do we compare float with zero without threshold?
     if (deltaSpeed != 0.0) {
-        acceleration = ((deltaSpeed / 3.6) / ((uint32_t)deltaTime / US_PER_SECOND_F));
+        acceleration = ((deltaSpeed / 3.6) / (deltaTime / US_PER_SECOND_F));
         if (direction) {
             //decceleration
             acceleration *= -1;
